restrict fake-xml-http-request dependency
fake-xml-http-request v2.1.1 introduced breaking changes to the handling
of the `response` property. In previous releases, `response` was not always
present, allowing fake-xml-http-response to interact with libraries such as
[whatwg-fetch](https://github.com/github/fetch/blob/96b37eb63ef8f2644f985cd77e7bd87b04e61ad4/fetch.js#L462)
that make decisions based on that behavior. Because pretender has a
caret dep on fake-xml-http-request, that change effectively broke
the previous major release (v2.x) of pretender because the PR #280 is
only available on the current major release (v3.x).

This PR simply changes the dependency on fake-xml-http-request from a
caret `^` to a tilde `~` in order to prevent that breaking change from
leaking through. If merged to a branch split from the last v2 commit and
released as v2.1.2, it would restore the previous behavior without
having any effect on the current major release.